requirements.txt
streamlit
openai
firecrawl
python-dotenv
nltk
